split item db commands into a separate module
